backend/requirements.txt
aiohttp==3.8.5
aiosignal==1.3.1
asgiref==3.7.2
async-timeout==4.0.3
attrs==23.1.0
autobahn==23.6.2
Automat==22.10.0
certifi==2023.7.22
cffi==1.16.0
channels==3.0.5
charset-normalizer==3.2.0
constantly==15.1.0
cryptography==41.0.4
Django==4.2.5
django-cors-headers==4.2.0
django-eventstream==4.5.1
django-grip==3.4.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
drf-spectacular==0.26.5
frozenlist==1.4.0
gripcontrol==4.2.0
hyperlink==21.0.0
idna==3.4
incremental==22.10.0
inflection==0.5.1
jsonschema==4.19.1
jsonschema-specifications==2023.7.1
MarkupSafe==2.1.3
multidict==6.0.4
openai==0.28.1
pubcontrol==3.5.0
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
PyJWT==2.8.0
pyOpenSSL==23.2.0
pytz==2023.3.post1
PyYAML==6.0.1
referencing==0.30.2
requests==2.31.0
rpds-py==0.10.3
service-identity==23.1.0
six==1.16.0
sqlparse==0.4.4
tqdm==4.66.1
Twisted==23.8.0
txaio==23.1.1
typing_extensions==4.8.0
uritemplate==4.1.1
urllib3==2.0.5
Werkzeug==2.3.7
yarl==1.9.2
zope.interface==6.0
